+++ HISTORY.md
@@ -59,6 +59,7 @@
 
 ### 2026-01-04
 * gui, objects, scenes: new "hide empty categories" setting
+* store: serialize "hide empty categories" setting in the project file
 
 ### 2026-01-03
 * gui, objects, threads, extensions, byob: refactored refreshing empty block category buttons